+++ src/interlegis/portalmodelo/policy/tests/test_portal_settings.py
@@ -163,6 +163,12 @@ class HomePropertiesTestCase(unittest.TestCase):
         home_action = self.actions.index_html
         self.assertEqual(home_action.url_expr, 'string:${portal_url}')
 
+    def test_login_action(self):
+        self.portal = self.layer['portal']
+        self.actions = api.portal.get_tool('portal_actions').site_actions
+        login_action = self.actions.login
+        self.assertEqual(login_action.url_expr, 'string:${globals_view/navigationRootUrl}/login}')
+
 
 class PloneBoardTestCase(unittest.TestCase):